Moved parts of clic into two new files
